requirements.txt
yfinance
requests
transformers
pandas
numpy
auto-sklearn
scikit-learn
matplotlib
seaborn
streamlit
python-dotenv